requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pandas==2.2.2
numpy==1.26.4
scipy==1.13.0
python-dateutil==2.9.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
azure-storage-blob==12.20.0
azure-identity==1.16.0
pyarrow==16.0.0
